requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backoff==2.2.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
dataclasses-json==0.5.7
defusedxml==0.7.1
deprecation==2.0.7
distlib==0.3.6
Django==4.2.4
django-allauth==0.54.0
django-bower==5.2.0
django-ckeditor==6.7.1
django-cors-headers==4.2.0
django-countries==7.5.1
django-htmx==1.16.0
django-paypal==2.0
django-sass-processor==1.2.2
django-scheduler==0.10.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
filelock==3.12.2
frozenlist==1.3.3
geographiclib==2.0
geopy==2.3.0
gpt-engineer==0.0.7
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-api-python-client==2.15.0
gunicorn==21.2.0
hyperlink==21.0.0
icalendar==5.0.7
identify==2.5.24
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
msgpack==1.0.5
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
openai==0.27.8
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.8.0
pluggy==1.2.0
pre-commit==3.3.3
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
pytest==7.3.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rudder-sdk-python==2.0.2
ruff==0.0.272
schedule==1.2.0
service-identity==23.1.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
termcolor==2.3.0
tomli==2.0.1
tqdm==4.65.0
Twisted==22.10.0
txaio==23.1.1
typer==0.9.0
typing-extensions==4.7.1
typing-inspect==0.9.0
Unipath==1.1
urllib3==2.0.3
virtualenv==20.24.0
yarl==1.9.2
zope.interface==6.0
